requirements.txt
transformers
python-dotenv
torch
datasets
nltk
soundfile
SentencePiece
pydub